Machine Learning System Engineering/lab_3/lab3/pyproject.toml
[tool.poetry]
name = "lab2 Training System"
version = "0.2.0"
description = ""
authors = ["James Winegar <jameswine@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
joblib = "^1.3.2"
scikit-learn = "1.3.0"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
fastapi-cache2 = "^0.1.9"
aioredis = "^2.0.1"
redis = "^5.0.1"
pytest-asyncio = "^0.21.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
